老师用户之间分隔开

+++ server/routers/student.js
@@ -15,10 +15,10 @@ router.use(async (ctx,next) => {
 router.post('/add',async ctx => {
     let {student_code,name,class_id,lesson_ids} = ctx.request.fields
     let row = await ctx.db.select('students_table','id',{
-        student_code
+        student_code,userId: ctx.session.userId
     })
     if(row.length > 0) {
-        ctx.body = {code:0,msg: '该学号已存在'}
+        ctx.body = {code:0,msg: '该学生已存在'}
         return
     }
     await ctx.db.insert('students_table',{
@@ -28,25 +28,25 @@ router.post('/add',async ctx => {
     ctx.body = {code:1,msg: 'OK'}
 })
 
-//添加学生
+//更新学生
 router.post('/update',async ctx => {
     let {id,student_code,name,lesson_ids} = ctx.request.fields
     await ctx.db.update('students_table',{
         student_code,name,lesson_ids
-    },{id})
+    },{id,userId: ctx.session.userId})
     ctx.body = {code:1,msg: 'OK'}
 })
 
 //删除学生
 router.post('/delete',async ctx => {
     let {id} = ctx.request.fields
-    await ctx.db.delete('students_table',{id})
+    await ctx.db.delete('students_table',{id,userId: ctx.session.userId})
     //删除点名记录
-    let records = await ctx.db.select('record_table','id,lesson_id',{student_id:id})
+    let records = await ctx.db.select('record_table','id,lesson_id',{student_id:id,userId: ctx.session.userId})
     for(let i=0;i<records.length;i++){
         let rec = records[i]
-        await ctx.db.delete('record_table',{id:rec.id})
-        await ctx.db.query(`UPDATE lesson_table SET attend_sum = attend_sum-1 WHERE id = ${rec.lesson_id}`)
+        await ctx.db.delete('record_table',{id:rec.id,userId: ctx.session.userId})
+        await ctx.db.query(`UPDATE lesson_table SET attend_sum = attend_sum-1 WHERE id = ${rec.lesson_id} AND userId = ${ctx.session.userId}`)
     }
     ctx.body = {code:1,msg: 'OK'}
 })
@@ -56,8 +56,8 @@ router.get('/class',async ctx => {
     let {pageSize,page} = ctx.query
     pageSize = pageSize || 15;
     page = page - 1 || 0
-    let total =  (await ctx.db.select('class_table','count(1)'))[0]['count(1)']
-    let list = await ctx.db.select('class_table','*','1=1',`${page*pageSize},${pageSize}`)
+    let total =  (await ctx.db.select('class_table','count(1)',{userId: ctx.session.userId}))[0]['count(1)']
+    let list = await ctx.db.select('class_table','*',{userId: ctx.session.userId},`${page*pageSize},${pageSize}`)
     ctx.body = {code: 1, msg: {
         total,
         list
@@ -67,13 +67,13 @@ router.get('/class',async ctx => {
 //学生列表
 router.get('/list',async ctx => {
     let {id} = ctx.query
-    let list = await ctx.db.select('students_table','*',{class_id:id})
+    let list = await ctx.db.select('students_table','*',{class_id:id,userId: ctx.session.userId})
     for(let i = 0;i < list.length;i++){
         let student = list[i]
         let lids = list[i].lesson_ids.split(',')
         let lnames = []
         for(let j = 0;j < lids.length;j++) {
-            let name = (await ctx.db.select('lesson_table','name',{id:lids[j]}))[0]
+            let name = (await ctx.db.select('lesson_table','name',{id:lids[j],userId: ctx.session.userId}))[0]
             name = name ? name.name : ''
             lnames.push(name)
         }
